Bit perfect sound driver and automatic Saxman patching!
I still don't know how I managed it, but not only do we now have the sound driver assembling perfectly, we also have it being inserted into the ROM AND have a solution to automatically patch the Saxman compressor by means of a modified version of Natsumi's DualPCM (the source repository of which will be published shortly).

Unfortunately, the result is not bit-perfect, but that is simply because there is not currently a standalone Saxman compressor that both does Sega-accurate compression and leaves the size off the compressed file (Sonic 2 AS' s2p2bin seems to the only implementation that exists currently). Currently, the driver compression is being done with Flamewing's saxcmp, but ideally we want something that is bit-perfect. For now though, this is a monumental achievement, and one that not even a few days ago I believed was months away, and, when I first started this project, wasn't even sure was possible!